created post page and linked to post content
