+++ pages/_app.js
@@ -1,5 +1,5 @@
 import { SessionProvider } from "next-auth/react"
-import Layout from "../components/Layout"
+import Layout from "../components/OLD-Layout"
 import '../styles/globals.css';
 import TimeAgo from 'javascript-time-ago'
 import en from 'javascript-time-ago/locale/en.json'
