+++ pages/api/posts.js
@@ -8,11 +8,17 @@ export default async function handler(req, res) {
     const session = await getServerSession(req,res,authOptions);
 
     if (req.method === 'GET') {
-        const posts = await Post.find()
+        const {id} = req.query;
+        if (id) {
+            const post = await Post.findById(id).populate('author');
+            res.json(post)
+        } else {
+            const posts = await Post.find()
             .populate('author')
             .sort({createdAt: -1})
             .exec()
-        res.json(posts)
+            res.json(posts)
+        }
     }
 
     if (req.method === 'POST') {
